backend/requirements.txt
Django>=5.0.6
psycopg2-binary>=2.9.7
python-dotenv>=1.0.1
django-redis>=5.4.0
celery>=5.3.6
cloudinary>=1.39.1
Pillow>=10.0.1
requests>=2.31.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
PyJWT>=2.8.0
pyotp>=2.9.0
typesense>=0.15.0